* Extract swiped events addition function, so iframe can use it.
